Abstract install and update env template generation in a shared class.
