+++ src/Command/GenerateEnvironmentCommand.php
@@ -2,10 +2,13 @@
 
 namespace Drupal\Console\Combawa\Command;
 
+use Drupal\Console\Combawa\Generator\EnvironmentUpdateGenerator;
 use Drupal\Console\Command\Shared\ConfirmationTrait;
 use Drupal\Console\Core\Command\Command;
-use Drupal\Console\Core\Utils\StringConverter;
 use Drupal\Console\Combawa\Generator\EnvironmentInstallGenerator;
+use Drupal\Console\Core\Utils\CountCodeLines;
+use Drupal\Console\Core\Utils\FileQueue;
+use Drupal\Console\Core\Utils\TwigRenderer;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -15,17 +18,31 @@ class GenerateEnvironmentCommand extends Command {
   use ConfirmationTrait;
 
   /**
-   * @var EnvironmentInstallGenerator
+   * @var \Drupal\Console\Combawa\EnvironmentGeneratorAbstract
    */
   protected $generator;
+  /**
+   * @var TwigRenderer
+   */
+  protected $renderer;
+  /**
+   * @var FileQueue
+   */
+  protected $fileQueue;
+  /**
+   * @var CountCodeLines
+   */
+  protected $countCodeLines;
 
   /**
    * ProfileCommand constructor.
    *
    * @param EnvironmentInstallGenerator $generator
    */
-  public function __construct(EnvironmentInstallGenerator $generator) {
-    $this->generator = $generator;
+  public function __construct(TwigRenderer $renderer, FileQueue $file_queue, CountCodeLines $count_code_lines) {
+    $this->renderer = $renderer;
+    $this->fileQueue = $file_queue;
+    $this->countCodeLines = $count_code_lines;
     parent::__construct();
   }
 
@@ -149,9 +166,30 @@ class GenerateEnvironmentCommand extends Command {
   }
 
   /**
+   * Helper to set generator attribute and its contextual dependencies.
+   */
+  protected function setGenerator() {
+    $build_mode = exec('/usr/bin/env composer config extra.combawa.build_mode  -d ' . $this->drupalFinder->getComposerRoot());
+    if ($build_mode == 'install') {
+      $generator = new EnvironmentInstallGenerator($this->drupalFinder);
+    }
+    elseif ($build_mode == 'update') {
+      $generator = new EnvironmentUpdateGenerator($this->drupalFinder);
+    }
+    else {
+      throw new \InvalidArgumentException(sprintf('Build mode "%s" is invalid (only install or update supported).', $build_mode));
+    }
+    $generator->setRenderer($this->renderer);
+    $generator->setFileQueue($this->fileQueue);
+    $generator->setCountCodeLines($this->countCodeLines);
+    $this->generator = $generator;
+  }
+
+  /**
    * {@inheritdoc}
    */
   protected function execute(InputInterface $input, OutputInterface $output) {
+    $this->setGenerator();
 
     $generateParams = [
       'app_root' => $this->generator->getCombawaRoot(),
